Fix pbr version when deploying wsgi
We recently added a check when deploying the keystone.wsgi script that would
allow it to be smart enough to handle the keystone stable/liberty codebase and
master. This change was using 2015 as the keystone version but for
stable/liberty this has changed to 8.

This patch makes it so that the wsgi script matches what PBR gives reports back
for stable/liberty deployments.